Drop chance is now a float

+++ Assets/_Scripts/UI/Inventory/Items/Item.cs
@@ -15,7 +15,7 @@ public class Item : ScriptableObject
     public int SellValue;
     public int Cost;
     public string Description;
-    [Range(0, 100)] public int DropChance;
+    [Range(0f, 100f)] public float DropChance;
     public bool IsStackable;
 
     public virtual void Use(Inventory _inventory, EquipmentManager _equipmentManager)
